requirements.txt
bcrypt==4.0.0
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Dropzone==1.6.0
Flask-Login==0.6.2
Flask-Session==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
pymongo==4.2.0
pytz==2022.2.1
qrcode==7.3.1
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
